pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]


[project]
name = "scvi-criticism"
version = "0.1.0"
description = "Evaluation metrics for scvi-tools models"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "Valeh Valiollah Pour Amiri"},
]
maintainers = [
    {name = "Valeh Valiollah Pour Amiri", email = "valehvpa@berkeley.edu"},
]
urls.Documentation = "https://scvi-criticism.readthedocs.io/"
urls.Source = "https://github.com/YosefLab/scvi-criticism"
urls.Home-page = "https://github.com/YosefLab/scvi-criticism"
dependencies = [
    "anndata",
    "scanpy",
    "numpy",
    "pandas",
    "scipy",
    "rich",
    "matplotlib",
    "seaborn",
    "sklearn",
    # for debug logging (referenced from the issue template)
    "session-info",
    "xarray",
    "scvi-tools",
    "sparse",
]

[project.optional-dependencies]
dev = [
    # CLI for bumping the version number
    "bump2version",
    "pre-commit",
    "twine>=4.0.2",
    "ipython",
]
doc = [
    "sphinx>=4",
    "sphinx-book-theme>=0.3.3",
    "myst-nb",
    "sphinxcontrib-bibtex>=1.0.0",
    "sphinx-autodoc-typehints",
    # For notebooks
    "ipykernel",
    "ipython",
    "sphinx-copybutton",
]
test = [
    "pytest",
    "pytest-cov",
]

[tool.hatch.build.targets.wheel]
packages = ['src/scvi_criticism']

[tool.coverage.run]
source = ["scvi_criticism"]
omit = [
    "**/test_*.py",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
xfail_strict = true
addopts = [
    "--import-mode=importlib",  # allow using test files with same name
]

[tool.isort]
include_trailing_comma = true
multi_line_output = 3
profile = "black"
skip_glob = ["docs/*"]

[tool.black]
line-length = 120
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.jupytext]
formats = "ipynb,md"

[tool.cruft]
skip = [
    "tests",
    "src/**/__init__.py",
    "src/**/basic.py",
    "docs/api.md",
    "docs/changelog.md",
    "docs/references.bib",
    "docs/references.md",
    "docs/notebooks/example.ipynb"
]
